refactor(library): extract native cover healing into a service
